security: address remaining LOW/INFO audit items from #307
In order from the PR #307 triage list:

1. release-path containment (Go + TS): the bytes at release-path are
   executed under sudo. Route the resolved path through
   safepath.Workspace (Go runStart) and a mirrored realpath
   trusted-roots check (TS start.ts) so a hostile env or symlink cannot
   point the agent binary at an arbitrary filesystem location.
   All in-repo workflows use workspace-relative `bin/coldstep`, which
   stays valid.

2. allow-file resource caps (Go + TS): bound the comma list to 64 files
   and each file to 8 MiB (mirrors maxGitHubEventJSONBytes posture).
   Go errors out (strict parse path); TS warns and skips (matches its
   existing warn-and-continue read posture).

3. GHES compatibility: postPRComment now derives its endpoint from
   GITHUB_API_URL (https-only, validated) instead of hardcoding
   api.github.com, so the Bearer token is not sent to public GitHub
   from an Enterprise Server runner.

4. Markdown-structure-aware truncate: when the PR-comment truncation
   cut lands inside a fenced code block, close the fence so the
   _(truncated)_ marker still renders as Markdown.

5. clampPerCPUSumToUint32: compare through int64 so the saturation
   logic is correct (and compiles) independent of platform int width
   instead of assuming 64-bit amd64/arm64.

Items 1-4 carry regression tests where unit-testable (file caps, API
base URL, fence repair). gofmt/vet clean; actioncli + safepath tests
pass; tsc --noEmit clean; dist rebuilt.

+++ internal/agent/agent_linux_policy_maps.go
@@ -210,15 +210,17 @@ func readSendpageObservedCount(objs *defend.DefendObjects) uint32 {
 // readUint32PerCPUArraySum down to uint32 with explicit saturation. The
 // per-cpu values are uint32 but summed into an int across the CPU set;
 // in practice the result fits, but gosec G115 wants an explicit bounded
-// conversion before the narrowing cast.
+// conversion before the narrowing cast. The comparison goes through int64
+// so the function is correct (and compiles) independent of the platform
+// int width rather than assuming the 64-bit int of amd64/arm64.
 func clampPerCPUSumToUint32(n int) uint32 {
 	if n <= 0 {
 		return 0
 	}
-	if n > math.MaxUint32 {
+	if int64(n) > math.MaxUint32 {
 		return math.MaxUint32
 	}
-	return uint32(n)
+	return uint32(n) // #nosec G115 -- saturation bounds checked above //nolint:gosec
 }
 
 // buildDefendAllowedPlan unifies the compile-and-merge sequence shared by the
